support --bento_only option in unit tests
* create separate init-tests script for bento tests
* separates bento unit testing files from other unit files
* adds bento-specific unit testing utilities
* updates defineElement() fns to accept a window attribute (useful for testing)
