+++ extensions/amp-embedly-card/1.0/web-component.js
@@ -8,15 +8,17 @@ import {
 
 /**
  * Registers `<bento-embedly-card> component to CustomElements registry
+ * @param {typeof globalThis=} win
  */
-export function defineElement() {
-  defineBentoElement(BENTO_TAG, BaseElement);
-  defineElementEmbedlyKey();
+export function defineElement(win) {
+  defineBentoElement(BENTO_TAG, BaseElement, win);
+  defineElementEmbedlyKey(win);
 }
 
 /**
  * Registers <bento-embedly-key>` component to CustomElements registry
+ * @param {typeof globalThis=} win
  */
-export function defineElementEmbedlyKey() {
-  defineBentoElement(EMBEDLY_KEY_BENTO_TAG, EmbedlyKeyBaseElement);
+export function defineElementEmbedlyKey(win) {
+  defineBentoElement(EMBEDLY_KEY_BENTO_TAG, EmbedlyKeyBaseElement, win);
 }
