+++ src/bento/util/unit-helpers.js
@@ -1,0 +1,9 @@
+/**
+ * @param {Window} win
+ * @param {string} css
+ */
+export function adoptStyles(win, css) {
+  const style = win.document.createElement('style');
+  style.textContent = `<style>${css}</style>`;
+  win.document.head.appendChild(style);
+}
